Move Kinepolis ValueObjects to separate subFolder
