+++ src/Kinepolis/Parser/KinepolisParser.php
@@ -6,8 +6,8 @@ namespace CultuurNet\UDB3\Kinepolis\Parser;
 
 use CultuurNet\UDB3\Event\EventThemeResolver;
 use CultuurNet\UDB3\Event\ValueObjects\LocationId;
-use CultuurNet\UDB3\Kinepolis\ParsedMovie;
-use CultuurNet\UDB3\Kinepolis\ParsedPriceForATheater;
+use CultuurNet\UDB3\Kinepolis\ValueObject\ParsedMovie;
+use CultuurNet\UDB3\Kinepolis\ValueObject\ParsedPriceForATheater;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\MultipleSubEventsCalendar;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\SingleSubEventCalendar;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\SubEvents;
