+++ tests/Kinepolis/Parser/KinepolisParserTest.php
@@ -7,8 +7,8 @@ namespace CultuurNet\UDB3\Kinepolis\Parser;
 use CultuurNet\UDB3\Event\EventThemeResolver;
 use CultuurNet\UDB3\Event\ValueObjects\LocationId;
 use CultuurNet\UDB3\Json;
-use CultuurNet\UDB3\Kinepolis\ParsedMovie;
-use CultuurNet\UDB3\Kinepolis\ParsedPriceForATheater;
+use CultuurNet\UDB3\Kinepolis\ValueObject\ParsedMovie;
+use CultuurNet\UDB3\Kinepolis\ValueObject\ParsedPriceForATheater;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\BookingAvailability;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\BookingAvailabilityType;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\DateRange;
